app/requirements.api.txt
gql[all]>=3.0.0a4
aiohttp
sqlalchemy
databases[postgresql]
redis
certifi
motor
python-snappy
python-dateutil
celery[all]
